requirements.txt
GitPython==3.1.44
molfeat==0.10.1
numpy==2.2.2
omegaconf==2.3.0
pandas==2.2.3
rdkit==2024.9.4
scikit_learn==1.6.1
scipy==1.15.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
wandb==0.19.4